test(data): green every CI leg
- `test_bbox_from_polygon.py` and `test_coordinate_convention.py` imported `torchvision.transforms.v2` at module scope, so every CI leg installing another extra failed collection outright with `ModuleNotFoundError: No module named 'torchvision'` — four errors before a single test ran, on all ten ubuntu `kornia` and `albumentations` jobs. Both now import behind `_TORCHVISION_AVAILABLE`, the flag the rest of the suite already guards on.
- The two modules are guarded differently because they need different things. Every case in `test_bbox_from_polygon.py` warps through a `torchvision` affine, so it takes a module-level `pytestmark`. Only one test in `test_coordinate_convention.py` does, so that one gets the skip and the rasterizer and writer tests keep running on every leg.
- That test's transforms are built by a lambda per case rather than constructed in the `parametrize` list: a decorator argument is evaluated at collection, so an unbuilt `T.RandomHorizontalFlip(...)` there would still touch `torchvision` on a leg that skips the test.
- `Quantize(levels=2).apply(image, None).max()` had its expected output written as `np.uint8(255)`, which is numpy 2's scalar repr. The minimum-pin leg resolves numpy 1.26.4 and prints `255`. Wrapped in `int(...)` so the example asserts the value rather than the repr of whichever numpy is installed.
- `test_generated_bytes_are_unchanged` is skipped below Pillow 11.1. Bisected, not guessed: `10.0.0`, `10.4.0` and `11.0.0` paint exactly one pixel of the whole matrix differently — sample 0 of `detection-shape-primitives-64` at `(55, 44)`, red on those and background from `11.1.0` on — and one pixel is a different hash. `pillow>=10` stays the package floor, because the library works there; it is this fixture that needs a fixed rasterizer, and regenerating below that Pillow now raises rather than writing one rasterizer's pixels into the file.
- Only that one check is gated. Measured across numpy 1.26.4/Pillow 10.0.0 against numpy 2.2.6/Pillow 12.1.1, no label digest moved while 20 of 25 pixel digests did, which is the quantized-label half of this snapshot confirmed on a second stack.
- The `FusedCompose` example under "Recovering a tighter box under rotation" in `tasks.md` opens on `transforms`, `image` and `polygon`, none of which it defines: it is a fragment showing the call shape, not a runnable example. `generate_doc_tests.py` converts every fenced Python block into a test and `phmdoctest` isolates each one, so it became `test_code_62` and failed with `NameError: name 'transforms' is not defined` on every CI job and every local run, as it had since before this branch. Marked with the `phmdoctest-skip` directive the generator already passes through, rather than made runnable: a working pipeline needs a concrete transform, and the only ones available come from kornia, which several CI legs install no extras for.
- `make test-matrix` passes every leg: albumentations 2819, kornia 2781, torchvision 2663, all 3576, minimum pins 3545.

---

+++ src/fuse_augmentations/data/degradations.py
@@ -23,8 +23,8 @@ Examples:
     >>> import numpy as np
     >>> from fuse_augmentations.data.degradations import Contrast, Quantize
     >>> image = np.full((4, 4, 3), 200, dtype=np.uint8)
-    >>> Quantize(levels=2).apply(image, None).max()
-    np.uint8(255)
+    >>> int(Quantize(levels=2).apply(image, None).max())
+    255
     >>> Contrast(factor=0.5).apply(image, None).shape
     (4, 4, 3)
